Support for error handling in AddParticipantsViewController. Refactoring of AddBotError Extension.
