+++ Wire-iOS/Sources/UserInterface/AddContacts/AddParticipantsViewController.swift
@@ -292,18 +292,28 @@ extension AddParticipantsViewController: SearchResultsViewControllerDelegate {
         // no-op
     }
     
+
     public func searchResultsViewController(_ searchResultsViewController: SearchResultsViewController, didTapOnSeviceUser user: ServiceUser) {
-        let serviceDetails = ServiceDetailViewController(serviceUser: user, variant: .light)
-        serviceDetails.destinationConversation = self.conversation
-        serviceDetails.completion = { [weak self] _ in
-            guard let `self` = self else {
-                return
-            }
-            self.dismiss(animated: true) {
-                self.delegate?.addParticipantsViewControllerDidCancel(self)
+        
+        let detail = ServiceDetailViewController(serviceUser: user, variant: .light)
+        detail.destinationConversation = self.conversation
+        detail.completion = { [weak self] result in
+            guard let `self` = self else { return }
+            
+            if let result = result {
+                switch result {
+                case .success( _):
+                    self.dismiss(animated: true, completion: {
+                        self.delegate?.addParticipantsViewController(self, didSelectUsers: [user as! ZMUser])
+                    })
+                case .failure(let error):
+                    error.displayAddBotError(in: detail)
+                }
             }
         }
-        self.present(serviceDetails.wrapInNavigationController(), animated: true, completion: nil)
+        
+        self.present(detail.wrapInNavigationController(), animated: true, completion: nil)
     }
+    
 }
